job_scheduler/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
astunparse==1.6.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
config==0.5.1
debugpy==1.8.1
decorator==5.1.1
dm-tree==0.1.8
executing==2.0.1
fastapi==0.115.12
filelock==3.18.0
Flask==2.1.2
flatbuffers==25.1.24
fsspec==2025.5.1
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.70.0
h11==0.16.0
h5py==3.12.1
huggingface-hub==0.32.4
idna==3.10
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mysql-connector-python==9.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.5
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.0
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==5.29.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.17.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.3
wcwidth==0.2.13
Werkzeug==2.1.2
wheel==0.43.0
wrapt==1.17.2
zipp==3.18.1